fix: extra refactoring
